Move selective logger to common.py
